feat(services): kitsu service
