+++ src/Commands/General/anime.ts
@@ -1,15 +1,16 @@
-import {
-	EmbedBuilder,
-	ApplicationCommandOptionType,
-	ChatInputCommandInteraction,
-} from 'discord.js';
+import { EmbedBuilder, ApplicationCommandOptionType, ChatInputCommandInteraction } from 'discord.js';
 import { Discord, Guard, Slash, SlashOption } from 'discordx';
 import { RateLimit, TIME_UNIT } from '@discordx/utilities';
-import { KitsuAPI } from '../../Utils/APIs/kitsuAPI.js';
+import { inject, injectable } from 'tsyringe';
+import { Kitsu } from '@Services';
 
 @Discord()
+@injectable()
 export class Anime {
 	private embed: EmbedBuilder;
+
+	// eslint-disable-next-line no-empty-function
+	constructor(@inject(Kitsu) private readonly kitsu: Kitsu) {}
 
 	@Slash({
 		description: 'Get info about an anime using Kitsu.io.',
@@ -31,47 +32,55 @@ export class Anime {
 			title: string,
 			interaction: ChatInputCommandInteraction,
 	) {
-		const kitsu = new KitsuAPI(interaction);
-		const anime = await kitsu.fetchAnime(title);
 		this.embed = new EmbedBuilder().setColor('Blurple').setTimestamp();
 
-		return interaction.reply({
-			embeds: [
-				this.embed
-					.setTitle(anime.titles.en_us)
-					.setThumbnail(anime.posterImage.original)
-					.setDescription(anime.synopsis)
-					.addFields(
-						{
-							name: 'Genres',
-							value: `> ${anime.genres}`,
-						},
-						{
-							name: 'Aired between',
-							value: `> <t:${anime.startDateUnix}> - <t:${anime.endDateUnix}>`,
-						},
-						{
-							name: 'Japanese Title',
-							value: `> ${anime.titles.ja_JP}`,
-							inline: true,
-						},
-						{
-							name: 'Status',
-							value: `> ${anime.status}`,
-							inline: true,
-						},
-						{
-							name: 'Average Rating',
-							value: `> ${anime.averageRating} / 100`,
-							inline: true,
-						},
-						{
-							name: 'Episodes',
-							value: `> ${anime.episodeCount}`,
-							inline: true,
-						},
+		await this.kitsu.fetchAnime(title).then((anime) => {
+			return interaction.reply({
+				embeds: [
+					this.embed
+						.setTitle(anime.titles.en_us)
+						.setThumbnail(anime.posterImage.original)
+						.setDescription(anime.synopsis)
+						.addFields(
+							{
+								name: 'Genres',
+								value: `> ${anime.genres}`,
+							},
+							{
+								name: 'Aired between',
+								value: `> <t:${anime.startDateUnix}> - <t:${anime.endDateUnix}>`,
+							},
+							{
+								name: 'Japanese Title',
+								value: `> ${anime.titles.ja_JP}`,
+								inline: true,
+							},
+							{
+								name: 'Status',
+								value: `> ${anime.status}`,
+								inline: true,
+							},
+							{
+								name: 'Average Rating',
+								value: `> ${anime.averageRating} / 100`,
+								inline: true,
+							},
+							{
+								name: 'Episodes',
+								value: `> ${anime.episodeCount}`,
+								inline: true,
+							},
+						),
+				],
+			});
+		}).catch(() => {
+			return interaction.reply({
+				embeds: [
+					this.embed.setDescription(
+						'🔹 | There was an error while fetching the information from the API.',
 					),
-			],
+				],
+			});
 		});
 	}
 }
