+++ src/Services/index.ts
@@ -1,3 +1,5 @@
 import { RAWG } from './RAWG';
+import { Kitsu } from './Kitsu';
+import { GoogleBooks } from './GoogleBooks';
 
-export { RAWG };
+export { RAWG, Kitsu, GoogleBooks };
